auth: Add concept of bk passwd update and policy
This adds the concept of a PasswordPolicy registration system, which
provides an extensible way of administering and configuring password
complexity policies.

It also adds a setPassword() method for the authentication backends which
will allow for the respective backend up update the password according to
whatever method is suitable for the respective backend (such as remote
updates for LDAP).